requirements.txt
tensorflow==2.16.1
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.10.7
pyyaml==6.0.1